artf2214
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@1733 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/engine/export/JRXlsExporter.java
@@ -162,13 +162,13 @@ public class JRXlsExporter extends JRXlsAbstractExporter
 
 	protected void setRowHeight(int rowIndex, int lastRowHeight)
 	{
-		row = sheet.getRow((short)rowIndex);		
+		row = sheet.getRow(rowIndex);		
 		if (row == null)
 		{
-			row = sheet.createRow((short)rowIndex);
+			row = sheet.createRow(rowIndex);
 		}
 		
-		row.setHeightInPoints((short)lastRowHeight);
+		row.setHeightInPoints(lastRowHeight);
 	}
 
 	protected void setCell(int colIndex, int rowIndex)
